kumo_dkim

Struct SignerBuilder

pub struct SignerBuilder { /* private fields */ }

Builder for the Signer

Implementations

impl SignerBuilder

pub fn new() -> Self

New builder

pub fn with_signed_headers( self, headers: impl IntoIterator<Item = impl Into<String>>, ) -> Result<Self, DKIMError>

Specify headers to be used in the DKIM signature The From: header is required.

pub fn with_over_signing(self, over_sign: bool) -> Self

Enable automatic over-signing. Each configured header in the list of headers to sign will be counted in the message, and it will be signed N+1 times so that the resultant signature will be proof against a replay attack that inserts an additional header of the same name.

pub fn with_private_key<K: Into<Arc<DkimPrivateKey>>>(self, key: K) -> Self

Specify the private key used to sign the email

pub fn with_selector(self, value: impl Into<String>) -> Self

Specify the private key used to sign the email

pub fn with_signing_domain(self, value: impl Into<String>) -> Self

Specify for which domain the email should be signed for

pub fn with_header_canonicalization(self, value: Type) -> Self

Specify the header canonicalization

pub fn with_body_canonicalization(self, value: Type) -> Self

Specify the body canonicalization

pub fn with_time(self, value: DateTime<Utc>) -> Self

Specify current time. Mostly used for testing

pub fn with_expiry(self, value: Duration) -> Self

Specify a expiry duration for the signature validity

pub fn build(self) -> Result<Signer, DKIMError>

Build an instance of the Signer Must be provided: signed_headers, private_key, selector and signing_domain.